Add Amulet of Yala to game spawn
In this commit, the 'Amulet of Yala' has been introduced to the game world. This new entity is spawned using the same system as player's spawn leveraging the MapBuilder and RandomNumberGenerator. The `AmuletOfYala` is officially a part of our game components and has distinct rendering properties. Updates have been made in main.rs, map_builder.rs, spawner.rs, and components.rs to facilitate this new feature.